requirements.txt
pyyaml==6.0.2
google-generativeai==0.8.3
openai==1.55.2
anthropic==0.39.0
pandas==2.1.3
python-dotenv==1.0.1
watchdog==6.0.0